requirements.txt
mlx
numpy
pyyaml
